refactor: restructure apollo graphql example
