+++ packages/rsocket-examples/package.json
@@ -15,7 +15,7 @@
     "start-client-server-request-response-websocket": "ts-node -r tsconfig-paths/register src/ClientServerRequestResponseExampleWebSocket.ts",
     "start-client-server-composite-metadata-route": "ts-node -r tsconfig-paths/register src/ClientServerCompositeMetadataRouteExample.ts",
     "start-client-server-rx-composite-metadata-route": "ts-node -r tsconfig-paths/register src/rxjs/ClientServerCompositeMetadataRouteExample.ts",
-    "start-client-server-apollo-graphql": "ts-node -r tsconfig-paths/register src/graphql/ApolloGraphQLRSocketLinkExample.ts"
+    "start-client-server-apollo-graphql": "ts-node -r tsconfig-paths/register src/graphql/apollo/example.ts"
   },
   "dependencies": {
     "@apollo/client": "^3.5.10",
